#33 Handle get requests & give available timeslots

Closes #33

See merge request courses/dit355/2023/student-teams/dit356-2023-12/toothtrek!26

+++ Middleware/bookings/src/main/java/com/toothtrek/bookings/repository/BookingRepository.java
@@ -7,7 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.toothtrek.bookings.entity.Booking;
 import com.toothtrek.bookings.entity.Patient;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
-    List<Booking> findByTimeslotId(Integer timeslotId);
+public interface BookingRepository extends JpaRepository<Booking, Long> {
+    List<Booking> findByTimeslotId(Long timeslotId);
+
     List<Patient> findByPatientId(String patientId);
 }
