+++ Middleware/bookings/src/main/java/com/toothtrek/bookings/repository/DentistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.toothtrek.bookings.entity.Dentist;
 
-public interface DentistRepository extends JpaRepository<Dentist, Integer> {
+public interface DentistRepository extends JpaRepository<Dentist, Long> {
 
 }
